Update VisioplanetPinItByImgSizeInShareSection.js
Since we're checking for document readiness, window.load may not be needed, and it's probably the culprit behind duplicacy and the cause is that both are called at the same time and by the time one call checks the true, the other is not done setting it. I wish we had something like synchronize in javascript as well.